// app/interface.tsx
export interface FormData {
  fullName: string;
  email: string;
  phone: string;
  experience: string;
  currentPosition: string;
  expectedSalary: string;
  coverLetter: string;
  portfolio: string;
  cv: File | null;
}

export type FormCareerErrors = Partial<Record<keyof FormData, string>>;
// ...
export interface JobData {
  title: string;
  department: string;
  location: string;
  type: string;
  experience: string;
  featured?: boolean;
  description: string;
  requirements: string[];
  salary: string;
  deadline: Date;
}

export interface JobPosition extends JobData {
  gradient: string;
  icon: React.ElementType;
}